Move OpenFoodNetwork::ProductsRenderer (lib) to ProductsRenderer (service) and refactor
